+++ docs/releases.md
@@ -1,8 +1,11 @@
 # Release Notes
 ## 0.1.1
 
-**2020-XX-XX**
+**2020-04-07**
+
+### Features
 - Support variables for `Operator` field
+- Add maxNumPVs option function to set a maximum number of PVs for a target
 
 ## 0.1.0
 
@@ -14,4 +17,3 @@
 - Data retrieval with data processing (See [Archiver Appliance User Guide](https://slacmshankar.github.io/epicsarchiver_docs/userguide.htm) for processing of data)
 - Using PV names for Grafana variables
 - Transform your data with processing functions
-